+++ tests/test_fmha_varlen.py
@@ -69,7 +69,7 @@ def _reset_everything():
     torch.manual_seed(42)
     torch.cuda.empty_cache()
 
-    # Hopper and Blackwell FMHA bwd don't have deterministic option.
+    # Hopper FMHA bwd doesn't have deterministic option.
     torch.use_deterministic_algorithms(False)
 
 
@@ -250,12 +250,13 @@ class FMHAVarlenTest(unittest.TestCase):
         heads_kv = heads_kv or heads
         head_dim_v = head_dim_v or head_dim
 
+        determinism = torch.are_deterministic_algorithms_enabled()
         logger.debug(
             f"Testing FMHA varlen ({backend}) vs {reference_backend}: {batch=}, {heads=}, {heads_kv=}, {head_dim=}, {head_dim_v=}, "
             f"{seqlens_Q_list=}, {seqlens_KV_list=}, {is_causal=}, {dtype=}, "
             f"{q_tile_size=}, {kv_tile_size=}, {run_persistent_kernel=}, {kernel_schedule=}"
             + (
-                f", {backward_q_tile_size=}, {backward_kv_tile_size=}, "
+                f",\n({determinism=}) {backward_q_tile_size=}, {backward_kv_tile_size=}, "
                 f"{backward_kv_splits=}, {backward_use_pt_reduction=}."
                 if test_backprop
                 else "."
@@ -508,6 +509,8 @@ class FMHAVarlenTest(unittest.TestCase):
         seqlens_KV_list,
         is_causal,
         heads_kv: Optional[int] = None,
+        determinism=False,
+        fail_case=False,
     ):
         torch.set_default_device("cuda")
 
@@ -525,6 +528,13 @@ class FMHAVarlenTest(unittest.TestCase):
             (torch.float8_e4m3fn, (1e-6, 1e-6), None),
             (torch.float8_e5m2, (1e-6, 1e-6), None),
         ]
+
+        if determinism:
+            ALLOWED_DTYPES = [
+                (torch.float16, (0, 0), (0, 0, 0)),
+                (torch.bfloat16, (0, 0), (0, 0, 0)),
+            ]
+            torch.use_deterministic_algorithms(not fail_case)
 
         for dtype, atol_fwd, atol_bwd in ALLOWED_DTYPES:
             dummy = torch.empty((1, 128, heads, head_dim), device="cuda", dtype=dtype)
@@ -565,6 +575,8 @@ class FMHAVarlenTest(unittest.TestCase):
                         run_persistent_kernel=run_persistent_kernel,
                     )
 
+        torch.use_deterministic_algorithms(False)
+
     def _test_varlen_backend(
         self,
         batch,
@@ -576,8 +588,12 @@ class FMHAVarlenTest(unittest.TestCase):
         backend,
         heads_kv=None,
         head_dim_v=None,
+        determinism: bool = False,
+        fail_case: bool = False,
     ):
         if backend == "cutlass-fmha":
+            assert not determinism
+            assert not fail_case
             assert heads_kv is None or heads_kv == heads
             return self._test_cutlass_fmha_varlen(
                 batch=batch,
@@ -589,6 +605,8 @@ class FMHAVarlenTest(unittest.TestCase):
                 is_causal=is_causal,
             )
         elif backend == "hopper-fmha":
+            assert not determinism
+            assert not fail_case
             assert (
                 head_dim_v is None or head_dim_v == head_dim
             ), "Hopper FMHA does not allow head_dim_v."
@@ -616,11 +634,15 @@ class FMHAVarlenTest(unittest.TestCase):
                 seqlens_Q_list=seqlens_Q_list,
                 seqlens_KV_list=seqlens_KV_list,
                 is_causal=is_causal,
+                determinism=determinism,
+                fail_case=fail_case,
             )
         else:
             raise NotImplementedError(f"Add {backend=} to tests.")
 
-    def _test_varlen_randsweep(self, backend, max_tests=1000):
+    def _test_varlen_randsweep(
+        self, backend, max_tests=1000, determinism: bool = False
+    ):
         max_qk = 2**17
         for i in range(max_tests):
             batch = random.choice(range(1, 12))
@@ -636,7 +658,7 @@ class FMHAVarlenTest(unittest.TestCase):
                 heads_choices = range(1, 4)
             else:
                 assert backend == "cutlass-fmha"
-                head_dim_choices = range(8, 256 + 1, 8)
+                head_dim_choices = range(8, 256 + 1, 8)  # type: ignore[assignment]
                 heads_choices = range(1, 4)
                 supports_dim_v = True
 
@@ -655,8 +677,8 @@ class FMHAVarlenTest(unittest.TestCase):
                 head_dim if not supports_dim_v else random.choice(head_dim_choices)
             )
 
-            seqlens_Q_list = []
-            seqlens_KV_list = []
+            seqlens_Q_list = []  # type: ignore[var-annotated]
+            seqlens_KV_list = []  # type: ignore[var-annotated]
             for i in range(batch):
                 max_q = min(2**12, max(max_qk - sum(seqlens_Q_list), 24))
                 max_k = min(2**12, max(max_qk - sum(seqlens_KV_list), 24))
@@ -676,6 +698,7 @@ class FMHAVarlenTest(unittest.TestCase):
                     seqlens_KV_list=seqlens_KV_list,
                     is_causal=is_causal,
                     backend=backend,
+                    determinism=determinism,
                 )
 
     @skip_if_libnatten_is_not_supported()
@@ -783,6 +806,47 @@ class FMHAVarlenTest(unittest.TestCase):
 
     @skip_if_libnatten_is_not_supported()
     @skip_if_blackwell_kernels_not_supported()
+    def test_cutlass_blackwell_varlen_determinism_fmha_fast(self):
+        problem_sizes = [
+            (
+                9,
+                4,
+                128,
+                [2669, 2240, 910, 2421, 3323, 34, 3308, 2867, 1401],
+                [2880, 1726, 1847, 1147, 3568, 3116, 661, 1739, 1146],
+            ),
+            (6, 1, 128, [128, 128, 135, 121, 128, 128], [128, 128, 135, 121, 128, 128]),
+            (5, 1, 128, [128, 128, 135, 128, 128], [128, 128, 135, 128, 128]),
+            (2, 1, 128, [135, 200], [128, 768]),
+            (2, 1, 128, [1024, 200], [128, 768]),
+            (2, 1, 128, [135, 200], [135, 768]),
+            (2, 1, 128, [1024, 200], [135, 768]),
+            (2, 1, 128, [1024, 256], [128, 768]),
+            (4, 1, 128, [1024, 8, 17, 2048], [10, 20, 512, 16]),
+            (3, 2, 128, [268, 1584, 1571], [2448, 4088, 1925]),
+            (2, 1, 128, [1024, 256], [512, 768]),
+        ]
+        for (
+            batch,
+            heads,
+            head_dim,
+            seqlens_Q_list,
+            seqlens_KV_list,
+        ) in problem_sizes:
+            for is_causal in [False, True]:
+                self._test_varlen_backend(
+                    batch=batch,
+                    heads=heads,
+                    head_dim=head_dim,
+                    seqlens_Q_list=seqlens_Q_list,
+                    seqlens_KV_list=seqlens_KV_list,
+                    is_causal=is_causal,
+                    backend="blackwell-fmha",
+                    determinism=True,
+                )
+
+    @skip_if_libnatten_is_not_supported()
+    @skip_if_blackwell_kernels_not_supported()
     def test_cutlass_blackwell_varlen_fmha_fast(self):
         problem_sizes = [
             (
@@ -827,6 +891,16 @@ class FMHAVarlenTest(unittest.TestCase):
     def test_cutlass_blackwell_varlen_fmha_extended(self):
         self._test_varlen_randsweep(
             backend="blackwell-fmha", max_tests=RAND_SWEEP_TESTS
+        )
+
+    @skip_if_not_running_extended_tests()
+    @skip_if_libnatten_is_not_supported()
+    @skip_if_blackwell_kernels_not_supported()
+    def test_cutlass_blackwell_varlen_fmha_determinism_extended(self):
+        self._test_varlen_randsweep(
+            backend="blackwell-fmha",
+            max_tests=max(2, RAND_SWEEP_TESTS // 100),
+            determinism=True,
         )